tests/inputdevice-keyboard.html

<!DOCTYPE html>
<title>InputDevice test for keyboard-triggered events</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>
<style>
#target {
  margin-bottom: 20px;
}
</style>

<h1>InputDevice - test for keyboard-triggered events</h1>
<p>Tab though the input fields below and activate the checkbox using nothing but a keyboard.
Spacebar is used to activate the checkbox in most browsers.  In Safari Option-Tab is necessary to
navigate to the checkbox.</p>
<div id="target" tabindex=1>
  <input>
  <input type=checkbox>
</div>

<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceValidator(sourceDevice) {
  assert_true(sourceDevice != null, "sourceDevice is set");
  assert_false(sourceDevice.firesTouchEvents, "sourcedevice.firesTouchEvents");
}

function onCompositionStart() {
  document.removeEventListener("compositionstart", onCompositionStart, true);
  runInputDeviceEventTests(["compositionstart", "compositionupdate", "compositionend"], inputDeviceValidator);
}

function runTests() {
  var events = ["click", "focus", "blur", "keydown", "keyup", "keypress"];

  // Mozilla doesn't support focusin/focusout: https://bugzilla.mozilla.org/show_bug.cgi?id=687787
  // There doesn't appear to be any good way to feature-detect for it.
  if (navigator.userAgent.indexOf("Firefox") == -1)
    events = events.concat(["focusin", "focusout"]);

  // Some browsers don't yet implement 'input' as a UIEvent, and many browsers don't yet support
  // 'beforeinput'
  if ("InputEvent" in window) {
    if ("oninput" in document.body)
      events.push("input");
    if ("onbeforeinput" in document.body)
      events.push("beforeinput");
  }

  if ("CompositionEvent" in window) {
    // Composition events may be generated (if we're using an IME).  Watch for them
    // and add more tests dynamically.
    document.addEventListener("compositionstart", onCompositionStart, true);
  }

  runInputDeviceEventTests(events, inputDeviceValidator);
}
</script>
